chore(opinionList): renaming to be consistent with other components
