+++ client/components/index.js
@@ -8,7 +8,7 @@ export { default as UserHome } from './user-home'
 export { Login, Signup } from './auth-form'
 export { default as Confirmation } from './confirmation'
 export { default as TransactionReceipt } from './transactionReceipt'
-export { default as OpinionList } from './opinion-list'
+export { default as OpinionList } from './opinionList'
 export { default as OpinionSelector } from './opinionSelector'
 export { default as CartForm } from './cartForm'
 export { default as CartFormTmpl } from './cartForm.tmpl'
