+++ client/routes.js
@@ -2,8 +2,7 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { withRouter, Route, Switch } from 'react-router-dom'
 import PropTypes from 'prop-types'
-import { Login, Signup, UserHome, Confirmation, CartForm } from './components'
-import OpinionList from './components/opinion-list'
+import { Login, Signup, UserHome, Confirmation, CartForm, OpinionList } from './components'
 import { me } from './store'
 /**
  * COMPONENT
